set terminal as interactive
