+++ dmsgpty/cli.go
@@ -3,9 +3,9 @@ package dmsgpty
 import (
 	"context"
 	"fmt"
-	"io"
 	"net"
 	"os"
+	"runtime"
 
 	"github.com/sirupsen/logrus"
 	"github.com/skycoin/skycoin/src/util/logging"
@@ -51,11 +51,13 @@ func (cli *CLI) StartLocalPty(ctx context.Context, cmd string, args ...string) e
 		return err
 	}
 
-	restore, err := cli.prepareStdin()
-	if err != nil {
-		return err
+	if runtime.GOOS != "Windows" {
+		restore, err := cli.prepareStdin()
+		if err != nil {
+			return err
+		}
+		defer restore()
 	}
-	defer restore()
 
 	return cli.servePty(ctx, ptyC, cmd, args)
 }
@@ -72,11 +74,13 @@ func (cli *CLI) StartRemotePty(ctx context.Context, rPK cipher.PubKey, rPort uin
 		return err
 	}
 
-	restore, err := cli.prepareStdin()
-	if err != nil {
-		return err
+	if runtime.GOOS != "Windows" {
+		restore, err := cli.prepareStdin()
+		if err != nil {
+			return err
+		}
+		defer restore()
 	}
-	defer restore()
 
 	return cli.servePty(ctx, ptyC, cmd, args)
 }
@@ -126,41 +130,3 @@ func (cli *CLI) prepareStdin() (restore func(), err error) {
 	return
 }
 
-// servePty serves a pty connection via the dmsgpty-host.
-func (cli *CLI) servePty(ctx context.Context, ptyC *PtyClient, cmd string, args []string) error {
-	ctx, cancel := context.WithCancel(ctx)
-	defer cancel()
-
-	cli.Log.
-		WithField("cmd", fmt.Sprint(append([]string{cmd}, args...))).
-		Infof("Executing...")
-
-	if err := ptyC.Start(cmd, args...); err != nil {
-		return fmt.Errorf("failed to start command on pty: %v", err)
-	}
-
-	// Window resize loop.
-	go func() {
-		defer cancel()
-		if err := ptyResizeLoop(ctx, ptyC); err != nil {
-			cli.Log.
-				WithError(err).
-				Warn("Window resize loop closed with error.")
-		}
-	}()
-
-	// Write loop.
-	go func() {
-		defer cancel()
-		_, _ = io.Copy(ptyC, os.Stdin) //nolint:errcheck
-	}()
-
-	// Read loop.
-	if _, err := io.Copy(os.Stdout, ptyC); err != nil {
-		cli.Log.
-			WithError(err).
-			Error("Read loop closed with error.")
-	}
-
-	return nil
-}
