+++ dmsgpty/cli_unix.go
@@ -5,6 +5,7 @@ package dmsgpty
 import (
 	"context"
 	"fmt"
+	"io"
 	"os"
 	"os/signal"
 	"syscall"
@@ -35,4 +36,43 @@ func ptyResizeLoop(ctx context.Context, ptyC *PtyClient) error {
 // getPtySize obtains the size of the local terminal.
 func getPtySize(t *os.File) (*pty.Winsize, error) {
 	return pty.GetsizeFull(t)
+}
+
+// servePty serves a pty connection via the dmsgpty-host.
+func (cli *CLI) servePty(ctx context.Context, ptyC *PtyClient, cmd string, args []string) error {
+	ctx, cancel := context.WithCancel(ctx)
+	defer cancel()
+
+	cli.Log.
+		WithField("cmd", fmt.Sprint(append([]string{cmd}, args...))).
+		Infof("Executing...")
+
+	if err := ptyC.Start(cmd, args...); err != nil {
+		return fmt.Errorf("failed to start command on pty: %v", err)
+	}
+
+	// Window resize loop.
+	go func() {
+		defer cancel()
+		if err := ptyResizeLoop(ctx, ptyC); err != nil {
+			cli.Log.
+				WithError(err).
+				Warn("Window resize loop closed with error.")
+		}
+	}()
+
+	// Write loop.
+	go func() {
+		defer cancel()
+		_, _ = io.Copy(ptyC, os.Stdin) //nolint:errcheck
+	}()
+
+	// Read loop.
+	if _, err := io.Copy(os.Stdout, ptyC); err != nil {
+		cli.Log.
+			WithError(err).
+			Error("Read loop closed with error.")
+	}
+
+	return nil
 }
